remove otto and tengo
